+++ src/main/java/com/theresourceroom/api/TheResourceRoomApp.java
@@ -1,13 +1,13 @@
 package com.theresourceroom.api;
 
-import com.theresourceroom.api.ServerConfiguration;
-
+import com.theresourceroom.api.Configuration;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
 
+
 // Main entry point for the application
-public class TheResourceRoomApp extends Application<ServerConfiguration> {
+public class TheResourceRoomApp extends Application<Configuration> {
 	
 	public static void main(String[] args) throws Exception {
         new TheResourceRoomApp().run(args);
@@ -18,17 +18,20 @@ public class TheResourceRoomApp extends Application<ServerConfiguration> {
         return "The Resource Room";
     }
 
+
     @Override
-    public void initialize(Bootstrap<ServerConfiguration> bootstrap) {
+    public void initialize(Bootstrap<Configuration> bootstrap) {
         // nothing to do yet
     }
     
     
     // This method is run via the main method to get the server up and running and register objects
     @Override
-    public void run(ServerConfiguration configuration, Environment environment) {
-
-        
+    public void run(Configuration configuration, Environment environment) {
+	    // nothing added yet
     }
+
+
+
 
 }