+++ src/main/java/com/theresourceroom/api/service/license/LicenseServiceWebImpl.java
@@ -3,6 +3,8 @@ package com.theresourceroom.api.service.license;
 import com.theresourceroom.api.dal.dao.LicenseDAO;
 import com.theresourceroom.api.models.License;
 
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 import java.util.Date;
 import java.util.List;
 
@@ -21,26 +23,44 @@ public class LicenseServiceWebImpl implements LicenseService {
 
     @Override
     public List<License> getLicensesForRoleFromSchool(String role, int school_id) {
+        role = role.strip().toUpperCase();
+        if(school_id < 1 || !role.matches("TEACHER") || !role.matches("ADMIN") || !role.matches("STUDENT") || !role.matches("PARENT")) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return licenseDAO.getLicensesForRoleFromSchool(role, school_id);
     }
 
     @Override
     public License getLicenseByNum(int num) {
+        if(num < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return licenseDAO.getLicenseByNum(num);
     }
 
     @Override
     public int createLicense(Date purchased, int purchased_by, String role) {
+        role = role.strip().toUpperCase();
+        if(purchased_by < 1 || !role.matches("TEACHER") || !role.matches("ADMIN") || !role.matches("STUDENT") || !role.matches("PARENT")) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return licenseDAO.createLicense(purchased, purchased_by, role);
     }
 
     @Override
     public License getUnusedLicenseForRole(int school, String role) {
+        role = role.strip().toUpperCase();
+        if(school < 1 || !role.matches("TEACHER") || !role.matches("ADMIN") || !role.matches("STUDENT") || !role.matches("PARENT")) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
         return licenseDAO.getUnusedLicenseForRole(school, role);
     }
 
     @Override
-    public int cancelLicense(int num) {
-        return licenseDAO.cancelLicense(num);
+    public boolean cancelLicense(int num) {
+        if(num < 1) {
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        }
+        return (licenseDAO.cancelLicense(num) == 1);
     }
 }