+++ src/main/java/com/theresourceroom/api/service/rating/RatingService.java
@@ -21,5 +21,5 @@ public interface RatingService {
 
     int updateRating(int rate, int id);
 
-    int deleteRating(int id);
+    boolean deleteRating(int id);
 }